.factory/skills/ai-info-stream-collector/ai-info-stream-collector-code/requirements.txt
 requests>=2.31.0
 beautifulsoup4>=4.12.0
 feedparser>=6.0.10
 pyyaml>=6.0.1
 apscheduler>=3.10.4
deep-translator>=1.11.4
 lxml>=4.9.3
 python-dateutil>=2.8.2
